ml/requirements.txt
transformers
torch
pillow
bitsandbytes
accelerate
python-dotenv